Дописан сбор объектов через фабрики

+++ Assets/Project/Scripts/DataBase/Data/EnemyData.cs
@@ -8,7 +8,7 @@ namespace Project.Scripts.DataBase.Data
     {
         [SerializeField] private string _id;
         [SerializeField] private string _name;
-        [SerializeField] private int _health;
+        [SerializeField] private float _health;
         [SerializeField] private int _damage;
         [SerializeField] private int _power;
         [SerializeField] private int _endurance;
@@ -17,7 +17,7 @@ namespace Project.Scripts.DataBase.Data
         
         public string Id => _id;
         public string Name => _name;
-        public int Health => _health;
+        public float Health => _health;
         public int Damage => _damage;
         public int Power => _power;
         public int Endurance => _endurance;
